Switch to cinder v2 endpoint API
nova-api filters cinder endpoints by "volume" and "volumev2", but needs
a further hint to actually use volumev2 from the catalog to use the v2
API. cinderclient also expects the endpoint to be volumev2 when asked to
use the v2 API. This commit adds the new endpoint in. The v1 endpoint is
left in place so that cinderclient continues to work without end users
needing to add the forced switch to v2, however we DO put that switch
into our stackrc.