Fix missing horizon local_settings in venv
There is an issue where when we are using source installs horizon does
not know where to find the local_settings.py. There are a few bad
options for fixing this, one of them is to link local_settings.py in to
the openstack_dashboard module in the venv. I think this is probably the
best bad option...