Make update-pip play to work by default
Right now if openstack.pypi_mirror is not defined then we end up with an
undefined variable error. The proper way to fix this would be to define
a default for this, but the play expects to be pointed at a devpi mirror
(see the appended +simple rather than /simple). This means that
pypi.python.org cannot be used by setting this variable.

Additionally, we have are checking a boolean (ansible version) but we
perform the same operation for both cases.